Make codechain-miner a framework
