+++ .gitignore
@@ -1,3 +1,4 @@
+Cargo.lock
 
 /target
 **/*.rs.bk
